serde_hkx::bytes
Module de

Modules

parser

Binary parser combinator

Structs

BytesDeserializer

Functions

from_bytes

Analyze as binary data of one file in order from hkx header.

from_bytes_with_opt

Analyze as binary data of one file in order from hkx header(with custom deserializer settings).

from_partial_bytes

Parse binary data as the type specified in the partial generics.

from_partial_bytes_with_opt

Parse binary data as the type specified in the partial generics with custom BytesDeserializer settings.