serde_hkx::bytes::serde
Module hkx_header

HKX Header Format Specification

The HKX header format is used for storing metadata information in HKX files. HKX files are binary files commonly used in video game development for storing animation and physics data. The header contains essential information about the structure and properties of the HKX file.

Size: 64bytes

Field Name	Description	Size (bytes)	Offset (bytes)

Magic0	First magic number (0x57E0E057)	4	0

Magic1	Second magic number (0x10C0C010)	4	4

UserTag	User-defined tag	4	8

FileVersion	Version of the file (LittleEndian e.g. 0x08 0x00 0x00 0x00)	4	12

PointerSize	Size of pointers in bytes (4 or 8)	1	16

Endian	Endianness of the file (0 for big-endian, 1 for little-endian)	1	17

PaddingOption	Padding option used in the file	1	18

BaseClass	Base class	1	19

SectionCount	Number of sections in the HKX file	4	20

ContentsSectionIndex	Index of the contents section within the file	4	24

ContentsSectionOffset	Offset of the contents section within the file	4	28

ContentsClassNameSectionIndex	Index of the contents class name section within the file	4	32

ContentsClassNameSectionOffset	Offset of the contents class name section within the file	4	36

ContentsVersionString	Version string of the contents (fixed-size string, 16 bytes)	16	40

Flags	Various flags used in the file	4	56

MaxPredicate	Maximum predicate value. None if -1.	2	60

SectionOffset	Section offset within the file. None if -1.	2	62

Paddings

If SectionOffset number is 16, read 64bytes header + an extra 16bytes as padding.

Field Name	Description	Size (bytes)	Offset (bytes)

Unk40	Unknown field (Hex offset: 40)	2	64

Unk42	Unknown field (Hex offset: 42)	2	66

Unk44	Unknown field (Hex offset: 44)	4	68

Unk48	Unknown field (Hex offset: 48)	4	72

Unk4C	Unknown field (Hex offset: 4C)	4	76

Structs

HkxHeader

The 64bytes HKX header contains metadata information about the HKX file.

Constants

SKYRIM_LE_ROW_HEADER

Skyrim LegendaryEdition(32bit) header binary

SKYRIM_SE_ROW_HEADER

Skyrim SpecialEdition(64bit) header binary