serde_hkx::bytes::serde::section_header

Struct SectionHeader

#[repr(C)]
pub struct SectionHeader {
    pub section_tag: [u8; 19],
    pub section_tag_separator: u8,
    pub absolute_data_start: u32,
    pub local_fixups_offset: u32,
    pub global_fixups_offset: u32,
    pub virtual_fixups_offset: u32,
    pub exports_offset: u32,
    pub imports_offset: u32,
    pub end_offset: u32,
}

The 48bytes each HKX section header contains metadata information about the HKX file.

For SkyrimSE, the bytes are arranged in the following order.

__classnames__ 48bytes

__types__ 48bytes

__data__ 48bytes

Note

This information is placed immediately after the Hkx header. (In some cases, padding is inserted in between.)

Depending on the havok version, there may be padding after the section header group. (at least not in SkyrimSE).

Fields
section_tag: [u8; 19]

Section name.

For SkyrimSE, the bytes are arranged in the following order.

__classnames__

__types__

__data__

Bytes Example

assert_eq!(
  [0x5F, 0x5F, 0x63, 0x6C, 0x61, 0x73, 0x73, 0x6E, 0x61, 0x6D, 0x65, 0x73, 0x5F, 0x5F, 0x00, 0x00, 0x00, 0x00, 0x00],
  *b"__classnames__\0\0\0\0\0"
);

section_tag_separator: u8

Always must be 0xFF

absolute_data_start: u32

Section start & fixup base offset.

Example of position

hkx_header.section_count:3 & hkx_header.section_offset:0 => 0x000000D0 bytes.

Calculation formula

Hkx header 64bytes + 48bytes * 3 sections = 208bytes == 0xD0(__classnames__ section abs)

local_fixups_offset: u32

Offset from absolute offset to local fixup map.

global_fixups_offset: u32

Offset from absolute offset to global fixup map.

virtual_fixups_offset: u32

Offset from absolute offset to virtual class fixup map.

exports_offset: u32

Unknown offset information.

Known information.

This value is the end position of the virtual_fixups_offset.

The exports, imports and end offsets are all the same value.

imports_offset: u32

Unknown offset information.

Known information.

This value is the end position of the virtual_fixups_offset.

The exports, imports and end offsets are all the same value.

end_offset: u32

Unknown offset information.

Known information.

This value is the end position of the virtual_fixups_offset.

The exports, imports and end offsets are all the same value.

Implementations

impl SectionHeader

pub const DATA_SECTION_HEADER_TAG: [u8; 19]

*b"__data__\0\0\0\0\0\0\0\0\0\0\0"

pub fn from_bytes<'a>( endian: Endianness, ) -> impl Parser<&'a [u8], Self, ContextError>

pub fn write_bytes<O>(&self, writer: impl WriteBytesExt) -> Result<()>
where O: ByteOrder,

Write section header to writer.

Errors

This method returns the same errors as Write::write_all.

pub fn write_classnames<O>( writer: impl WriteBytesExt, section_offset: i16, section_end_abs: u32, ) -> Result<()>
where O: ByteOrder,

Create new __classnames__ section header

section_offset: usually 0xff(ver. hk2010), this case padding is none.

Errors

This function will return the first error of non-[ErrorKind::Interrupted] kind that write returns.

pub fn write_types<O>(writer: impl WriteBytesExt, abs_offset: u32) -> Result<()>
where O: ByteOrder,

Write __types__ section header.

Errors

This function will return the first error of non-[ErrorKind::Interrupted] kind that write returns.