serde_hkx::errors::ser

Enum Error

pub enum Error {
    Message {
        msg: String,
    },
    InvalidEndian {
        invalid: u8,
        location: Location,
    },
    UnsupportedPtrSize {
        invalid: u8,
        location: Location,
    },
    OverflowSubtractAbs {
        position: u32,
        abs_data_offset: u32,
    },
    MissingGlobalFixupClass {
        ptr: Pointer,
        location: Location,
    },
    MissingClassInClassnamesSection {
        class_name: &'static str,
        location: Location,
    },
    NotFoundPointedPosition,
    Utf8Error {
        source: Utf8Error,
        location: Location,
    },
    NulError {
        source: NulError,
        location: Location,
    },
    IoError {
        source: Error,
        location: Location,
    },
}

Serialize error

Variants

Message

{msg}

Fields

msg: String

Error message

InvalidEndian

Only 0 (big) or 1 (little) can be specified for the header endian. But got {invalid}

Fields

invalid: u8

location: Location

error location

UnsupportedPtrSize

The only supported pointer sizes are 4 and 8. But got {invalid}

Fields

invalid: u8

location: Location

error location

OverflowSubtractAbs

Relative position cannot be obtained because abs is larger than {position}. This indicates that the value of absolute_data_offset({abs_data_offset}) in the header is wrong.

Fields

position: u32

abs_data_offset: u32

MissingGlobalFixupClass

Missing global fixup class: {ptr}

Fields

ptr: Pointer

missing global fixup class ptr(e.g. #0050)

location: Location

error location

MissingClassInClassnamesSection

The constructor class for virtual_fixup did not exist in the class in the __classnames__ section written.: {class_name}

Fields

class_name: &'static str

location: Location

error location

NotFoundPointedPosition

Not found where to write ptr. This could be an incorrect value inside hkx or a mistake by the library implementor.

Utf8Error

Invalid utf8 error

Fields

source: Utf8Error

Invalid utf8 error

location: Location

error location

NulError

Contain null bytes in a string error

Fields

source: NulError

Contain null bytes in a string error

location: Location

error location

IoError

std io error.

Fields

source: Error

I/O Error

location: Location

error location