serde_hkx::tree

Trait HavokTree

pub trait HavokTree {
    type Error;

    // Required methods
    fn tree_for_bytes(&mut self) -> Result<String, Self::Error>;
    fn unreferenced_indexes(&self) -> Result<Vec<usize>, Self::Error>;
}

Trait to create ptr dependencies tree

Required Associated Types

type Error

Required Methods

fn tree_for_bytes(&mut self) -> Result<String, Self::Error>

Tree of the order in which to serialize as binary data.

Errors

If hkRootLevelContainer is missing.

fn unreferenced_indexes(&self) -> Result<Vec<usize>, Self::Error>

Returns class indexes that are not reachable from hkRootLevelContainer.

These classes are not referenced by any dependency pointer and can be safely removed before serialization if they are not required externally.

Errors

If hkRootLevelContainer is missing.