serde_hkx::xml::de::parser
Module tag

XML tag parsers

Functions

attr_string

Parses a xml attribute string(surrounded double quotes), e.g. "string"

class_start_tag

Parses the array start tag (e.g. <hkobject name="#0010" class="hkbProjectData" signature="0x13a39ba7">)

end_tag

Parses the end tag </tag>

field_start_close_tag

Parses the field of class start closing tag > or numelements="0">

field_start_open_tag

Parses the field of class start opening tag <hkparam name=

number_in_string

Parses a number inside a string, e.g., "64"

start_tag

Parses the start tag <tag>